[FIX] l10n_es_pos: set simplifed partner only when we will invoice
Currently we also set the simplified partner directly on pos orders even
in case we do not invoice the pos order directly.
This is unnecessary; we only need to set a partner in case we invoice.
(Since we need a partner to put on the invoice.)

After this commit we only set the simplified partner on pos orders
that will be invoiced as simplified invoice.
(This happens automatically in case a simplified invoice journal is
set in the settings; see field `pos_l10n_es_simplified_invoice_journal_id`)

part of
task-3745982

X-original-commit: eba25b0f403380b8cd5b0e23ec252dae050e5beb
Part-of: odoo/odoo#219953
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>