[FIX] mass_mailing: give fallback value when user schedule a mail without body
This traceback occurs when the user schedules a mail without mail Body in
email marketing.

To reproduce this issue:
1) Install "mass_mail"
2) Create new mailings with 'contact_list_ids' and make sure to give
   the value for preview in the settings page.
3) Don't select any mail Body templates.
4) Schedule the record with the previous day.
5) An Error was encountered in the terminal.

Error:- "TypeError: expected string or bytes-like object"

When the user schedules a mail  corn job runs, in which "action_send_mail"
triggers.
In that method, while composing values "body" is used to get values from
mail.body_html.

https://github.com/odoo/odoo/blob/e37c393d415d686584487f9ad92f062279504cf8/addons/mass_mailing/models/mailing.py#L1043-L1046

Because when the user doesn't select the body template its value will be
false which leads to the above traceback,
as the re.search method.

https://github.com/odoo/odoo/blob/e37c393d415d686584487f9ad92f062279504cf8/odoo/tools/mail.py#L500

This commit will resolve the issue by giving a fallback value of an empty string
when the user doesn't select the body template.

sentry-5389317834

closes odoo/odoo#205150

X-original-commit: f65f9934fec9f6583756d1aa7ab967c8c54e8260
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>