[ADD] sms_twilio: new SMS provider
Current situation

We provide a SMS service for our clients through an IAP service. I
has almost no annoying configuration, the Odoo database sends data
to an IAP proxy, that sends the SMS to our providers for them. It
is easy to use and user friendly.

Issue

For this to work, Odoo "registers" itself to our providers in behalf
of our clients. The problem with Mobile operators in many countries
it is often asked for the clients to register themselves and submit
an authorization request to be able to send SMS on their network.

As such, Odoo is not supposed to register itself in place of their
customers. In some countries like USA, Australia, UK and soon France
this service is not working well or even at all.

Solution

Registration process is fastidious, and on top of them it changes
from country to country. We do not want to start implementing
registration processes by countries just for SMS sending.

For that reason, we will start to provide a new module where we
remove the IAP part and connect directly to a new provider: Twilio.
Clients will do all the registration themselves directly with the
provider, and setup their account in Odoo to send those SMS.
Twilio is a well known provider with extensive coverage, ideal
for a global solution like Odoo.

Stable target

This module targets 17.0 which is going to be the oldest supported
version soonish. It is going to be forward ported until master. It
should be stable compliant and should not break any flow involving
IAP contact.

Bridge support for SMS marketing and therefore marketing automation
is provided.

Task-4658352

X-original-commit: aa04aebff05b4c57668808376ef5ef4b7745d8fb
Part-of: odoo/odoo#224898
Related: odoo/enterprise#93569
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Co-authored-by: Thibault Delavallee <tde@odoo.com>